+++ tests/sim/status_warehouse/container/carousel/test_builtin_carousel.py
@@ -1,18 +1,20 @@
 import copy
 from unittest import TestCase
 
-from src.status_warehouse.container.carousel import Carousel, CarouselInfo
+from src.status_warehouse.container.carousel import Carousel
 from src.warehouse import Warehouse
+from src.warehouse_configuration_singleton import CarouselConfiguration
 
 
 class TestBuiltinCarousel(TestCase):
     def setUp(self):
         self.warehouse = Warehouse()
-        self.carousel_config = CarouselInfo (
+        self.carousel_config = CarouselConfiguration(
             bay_height = 150,
             buffer_height = 150,
             x_offset = 125,
-            width = 250
+            width = 250,
+            hole_height=150
         )
         self.carousel = Carousel(self.carousel_config, self.warehouse)
 
@@ -41,11 +43,12 @@ class TestBuiltinCarousel(TestCase):
         # arrange
         carousel_1 = self.carousel
         carousel_2 = Carousel(
-            CarouselInfo(
+            CarouselConfiguration(
                 bay_height=self.carousel_config.bay_height + 10,
                 buffer_height=self.carousel_config.buffer_height + 10,
                 x_offset=self.carousel_config.x_offset + 25,
-                width=self.carousel_config.width + 10
+                width=self.carousel_config.width + 10,
+                hole_height=self.carousel_config.hole_height + 10
             ),
             Warehouse())