+++ tests/sim/status_warehouse/container/column/test_builtin_column.py
@@ -1,14 +1,15 @@
 import copy
 from unittest import TestCase
 
-from src.status_warehouse.container.column import Column, ColumnInfo
+from src.status_warehouse.container.column import Column
 from src.warehouse import Warehouse
+from src.warehouse_configuration_singleton import ColumnConfiguration
 
 
 class TestBuiltinColumn(TestCase):
     def setUp(self):
         self.warehouse = Warehouse()
-        self.column_config = ColumnInfo(
+        self.column_config = ColumnConfiguration(
             height = 325,
             x_offset = 125,
             width = 250,
@@ -40,7 +41,7 @@ class TestBuiltinColumn(TestCase):
     def test_hash(self):
         # arrange
         column_1 = self.column
-        column_2 = Column(ColumnInfo(
+        column_2 = Column(ColumnConfiguration(
             height = 350,
             x_offset = 150,
             width = 275,