+++ tests/sim/status_warehouse/container/tray_container/test_builtin_tray_container.py
@@ -1,8 +1,9 @@
 from unittest import TestCase
 
-from src.status_warehouse.container.column import Column, ColumnInfo
+from src.status_warehouse.container.column import Column
 from src.status_warehouse.container.tray_container import TrayContainer
 from src.warehouse import Warehouse
+from src.warehouse_configuration_singleton import ColumnConfiguration
 
 
 class TestBuiltinTrayContainer(TestCase):
@@ -51,7 +52,7 @@ class TestBuiltinTrayContainer(TestCase):
     def test_hash(self):
         # arrange
         tray_container_1 = self.tray_container
-        tray_container_2 = Column(ColumnInfo(
+        tray_container_2 = Column(ColumnConfiguration(
             height = 5 * 25,
             x_offset = 150,
             width = 200,